Move parsing logic to own module
